requirements.txt
Flask==2.3.3
google-cloud-storage==2.10.0
gunicorn==21.2.0
Werkzeug==2.3.7
google-generativeai==0.8.2
pyrebase4==4.7.1
firebase-admin==6.2.0